AP-5203 Update test to extract meta data with column header map
